test(bridge): share one capturing logger across bridge test modules
Copilot: log::set_logger is process-global and only succeeds once —
this branch's references tests and main's response tests each installed
their own capturing logger, so whichever module ran second panicked.
(The failure was masked locally by a corrupt incremental target dir
that silently dropped the module from the test binary; reproduced on a
fresh target, 3 tests red.) Both modules now share a single
crate-level test_logging::captured_warnings_for with a serialize-lock,
and the transform doc comment reflects the
io::Result<Option<Vec<Location>>> contract.
